healthcare_backend/requirements.txt
django-cors-headers==4.5.0
python-dotenv==1.0.1
